first half of the forest change widgets
