+++ app/javascript/components/widgets/forest-change/fao-reforest/index.js
@@ -1,7 +1,48 @@
-import Component from 'components/widgets/components/widget-numbered-list';
-import getData from './actions';
-import getProps from './selectors';
-import config from './config';
-import settings from './settings';
+import { getFAOExtent } from 'services/forest-data';
 
-export { getData, getProps, Component, config, settings };
+import getWidgetProps from './selectors';
+
+export default {
+  widget: 'faoReforest',
+  title: {
+    global: 'Global FAO reforestation',
+    initial: 'FAO reforestation in {location}'
+  },
+  categories: ['forest-change'],
+  types: ['global', 'country'],
+  admins: ['global', 'adm0'],
+  settingsConfig: [
+    {
+      key: 'period',
+      label: 'period',
+      type: 'select'
+    }
+  ],
+  chartType: 'rankedList',
+  dataType: 'fao',
+  metaKey: 'widget_rate_of_restoration_fao',
+  sortOrder: {
+    forestChange: 8
+  },
+  colors: 'gain',
+  sentences: {
+    globalInitial:
+      'According to the FAO, the {location} rate of reforestation in {year} was {rate} per year.',
+    initial:
+      'According to the FAO, the rate of reforestation in {location} was {rate} per year in {year}.',
+    noReforest: 'No reforestation data in {location}.'
+  },
+  settings: {
+    period: 2010,
+    unit: 'ha/year',
+    pageSize: 5,
+    page: 0
+  },
+  getData: params =>
+    getFAOExtent({ ...params }).then(response => {
+      const data = response.data.rows;
+      const hasCountryData = (data.length && data.find(d => d.iso)) || null;
+      return hasCountryData ? data : {};
+    }),
+  getWidgetProps
+};
