--- app/javascript/components/widgets/forest-change/fao-reforest/settings.js
@@ -1,6 +1,0 @@
-export default {
-  period: 2010,
-  unit: 'ha/year',
-  pageSize: 5,
-  page: 0
-};
