--- app/javascript/components/widgets/forest-change/fires-alerts/actions.js
@@ -1,10 +1,0 @@
-import { fetchFiresAlerts, fetchFiresLatest } from 'services/alerts';
-import axios from 'axios';
-
-export default ({ params }) =>
-  axios.all([fetchFiresAlerts(params), fetchFiresLatest(params)]).then(
-    axios.spread((alerts, latest) => {
-      const { data } = alerts.data;
-      return { alerts: data, latest } || {};
-    })
-  );
