--- app/javascript/components/widgets/forest-change/fires-alerts/settings.js
@@ -1,7 +1,0 @@
-export default {
-  period: 'week',
-  weeks: 13,
-  dataset: 'VIIRS',
-  layerStartDate: null,
-  layerEndDate: null
-};
