+++ app/javascript/components/widgets/forest-change/fires-ranked/index.js
@@ -1,7 +1,52 @@
-import Component from 'components/widgets/components/widget-numbered-list';
-import getData from './actions';
-import getProps from './selectors';
-import config from './config';
-import settings from './settings';
+import axios from 'axios';
+import { fetchFiresAlertsGrouped, fetchFiresLatest } from 'services/alerts';
 
-export { getData, getProps, Component, config, settings };
+import getWidgetProps from './selectors';
+
+export default {
+  widget: 'firesRanked',
+  title: 'Regions with the most fire Alerts in {location}',
+  categories: ['forest-change'],
+  types: ['country'],
+  admins: ['adm0', 'adm1'],
+  settingsConfig: [
+    {
+      key: 'weeks',
+      label: 'weeks',
+      type: 'select',
+      whitelist: [13, 26, 52],
+      noSort: true
+    }
+  ],
+  chartType: 'rankedList',
+  metaKey: 'widget_fire_ranking',
+  colors: 'fires',
+  sortOrder: {
+    summary: 6,
+    forestChange: 10
+  },
+  sentences: {
+    initial:
+      'In the last {timeframe} in {location}, the region with the most fires burning was {topRegion}, with {topRegionCount} fire alerts, representing {topRegionPerc} of total alerts detected.',
+    withInd:
+      'In the last {timeframe} in {location}, the region with the most fires burning within {indicator} was {topRegion}, with {topRegionCount} fire alerts, representing {topRegionPerc} of total alerts detected.'
+  },
+  settings: {
+    unit: '%',
+    pageSize: 5,
+    page: 0,
+    period: 'week',
+    weeks: 13,
+    dataset: 'VIIRS',
+    layerStartDate: null,
+    layerEndDate: null
+  },
+  getData: params =>
+    axios.all([fetchFiresAlertsGrouped(params), fetchFiresLatest(params)]).then(
+      axios.spread((alerts, latest) => {
+        const { data } = alerts.data;
+        return { alerts: data, latest: latest.attributes.updatedAt } || {};
+      })
+    ),
+  getWidgetProps
+};
