--- app/javascript/components/widgets/forest-change/fires-ranked/settings.js
@@ -1,10 +1,0 @@
-export default {
-  unit: '%',
-  pageSize: 5,
-  page: 0,
-  period: 'week',
-  weeks: 13,
-  dataset: 'VIIRS',
-  layerStartDate: null,
-  layerEndDate: null
-};
