--- app/javascript/components/widgets/forest-change/glad-ranked/actions.js
@@ -1,27 +1,0 @@
-import { fetchGladIntersectionAlerts, fetchGLADLatest } from 'services/alerts';
-import { getExtentGrouped } from 'services/forest-data';
-import axios from 'axios';
-
-export default ({ params }) =>
-  axios
-    .all([
-      fetchGladIntersectionAlerts(params),
-      fetchGLADLatest(params),
-      getExtentGrouped(params)
-    ])
-    .then(
-      axios.spread((alerts, latest, extent) => {
-        const { data } = alerts.data;
-        const areas = extent.data.data;
-        const latestDate = latest.attributes && latest.attributes.updatedAt;
-
-        return data && extent && latest
-          ? {
-            alerts: data,
-            extent: areas,
-            latest: latestDate,
-            settings: { latestDate }
-          }
-          : {};
-      })
-    );
