--- app/javascript/components/widgets/forest-change/glad-ranked/settings.js
@@ -1,9 +1,0 @@
-export default {
-  threshold: 30,
-  extentYear: 2010,
-  unit: '%',
-  weeks: 4,
-  pageSize: 5,
-  page: 0,
-  ifl: 2016
-};
