--- app/javascript/components/widgets/forest-change/tree-cover-gain/settings.js
@@ -1,7 +1,0 @@
-export default {
-  threshold: 50,
-  unit: 'ha',
-  pageSize: 5,
-  page: 0,
-  ifl: 2000
-};
